brute force solution
in this implementation we check all possible combinations of coins that
match the expected change ammount.
It is possible to do this implementation in a bit more efficient
traversing and combination ordering. However, the benefits might only
apply to large trees(large number of coins), which in this case would not be required.